:building_construction: moved processPayment method to parent class
